Super basic github action
These are the two commands I've been running locally to make sure
everything is building and the tests are passing. I've just got an
old Flipper wifi dev board plugged in and the build is succeeding
which makes me think there's probably a bunch of stuff that isn't
being exercised by the build/compile step at the moment. Happy to
iterate on this and make it actually useful as a guard. I left both
steps as continue-on-error in the meantime.